* fixes, seems like all working fine. tomorrow updating docs
git-svn-id: file:///usr/local/src/otland/tfssvnrepo/trunk@266 c161f52b-5830-4e7a-bee0-5dea2c17262e

+++ luascript.cpp
@@ -7531,7 +7531,7 @@ int32_t LuaScriptInterface::luaAddEvent(lua_State* L)
 
 	eventDesc.parameters = params;
 
-	uint32_t delay = std::max((uint32_t)100, popNumber(L));
+	uint32_t delay = std::max((uint64_t)100, popNumber(L));
 	eventDesc.function = luaL_ref(L, LUA_REGISTRYINDEX);
 
 	eventDesc.scriptId = env->getScriptId();
